api/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
openmeteo_requests==1.3.0
pyodbc==5.1.0
python-dotenv==1.0.1
requests_cache==1.2.1
retry_requests==2.0.0
gunicorn==23.0.0
colorlog==6.9.0
msal==1.26.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0